# docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nyamby API",
    "version": "1.0.0",
    "description": "API spec untuk MVP Nyamby — Next.js Route Handlers. Base URL: /api",
    "contact": {
      "name": "Satria Romanda"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local Development" },
    { "url": "https://nyamby.vercel.app/api", "description": "Production (Vercel)" }
  ],
...
  "paths": {
...
    "/talent/onboarding": {
      "post": {
        "tags": ["Talent"],
        "summary": "Simpan onboarding form talenta",
        "description": "Menyimpan data profil talenta dan **otomatis men-trigger AI Skill Gap Analysis** setelah berhasil.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["bio", "category", "skills"],
                "properties": {
                  "bio": { "type": "string", "example": "Web developer dengan fokus React dan Next.js" },
                  "category": { "type": "string", "enum": ["web_dev", "graphic_designer"] },
                  "rate_per_hour": { "type": "number", "example": 75000 },
                  "rate_per_project": { "type": "number", "example": 2000000 },
                  "availability": { "type": "string", "enum": ["available", "busy", "unavailable"] },
                  "location": { "type": "string", "example": "Bandung" },
                  "portfolio_url": { "type": "string", "format": "uri", "example": "https://github.com/raka" },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "skill_id": { "type": "string", "format": "uuid" },
                        "level": { "type": "string", "enum": ["beginner", "intermediate", "expert"] }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profil disimpan + AI Skill Gap triggered",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Profil berhasil dibuat. AI sedang menganalisis skill gap kamu.",
                  "data": { "talent_profile_id": "uuid", "skill_gap_analysis_id": "uuid" }
                }
              }
            }
          }
        }
      }
    },
...
    "/talent/profile": {
      "patch": {
        "tags": ["Talent"],
        "summary": "Update profil talenta",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Semua field opsional — hanya kirim yang ingin diupdate",
                "properties": {
                  "bio": { "type": "string" },
                  "rate_per_hour": { "type": "number" },
                  "availability": { "type": "string", "enum": ["available", "busy", "unavailable"] },
                  "skills": { "type": "array" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profil berhasil diupdate" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/talent/dashboard": {
      "get": {
        "tags": ["Talent"],
        "summary": "Data dashboard talenta",
        "description": "Mengembalikan job rekomendasi (dari matching) + skill gap insight terbaru.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "skill_gap": {
                      "recommendations": [
                        { "skill_name": "TypeScript", "priority": "high", "reason": "Muncul di 23 job aktif", "estimated_impact": "Rate naik 30%" }
                      ],
                      "summary": "Skill kamu sudah solid untuk entry level..."
                    },
                    "recommended_jobs": [
                      { "job_id": "uuid", "title": "Landing Page UMKM", "match_score": 87, "budget_max": 2000000, "deadline": "2026-06-30" }
                    ],
                    "active_jobs": [
                      { "job_id": "uuid", "title": "Brand Identity", "status": "in_progress" }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/talent/skill-gap": {
      "get": {
        "tags": ["Talent"],
        "summary": "Ambil hasil AI Skill Gap Analysis terbaru",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "generated_at": "2026-05-28T10:00:00Z",
                    "recommendations": [
                      { "skill_name": "TypeScript", "priority": "high", "reason": "...", "estimated_impact": "..." },
                      { "skill_name": "Jest / Testing", "priority": "medium", "reason": "...", "estimated_impact": "..." },
                      { "skill_name": "Docker", "priority": "low", "reason": "...", "estimated_impact": "..." }
                    ],
                    "summary": "..."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token dari NextAuth.js. Sertakan di header: Authorization: Bearer <token>"
      }
    }
  }
}